removing ableist language
We should remove wording from the Handbook that is a slur. Especially in discussion of the Code of Conduct and really especially in a context of teaching instructors how to talk about a code of conduct if they are unaccustomed to doing so.

I think "clueless" captures what we mean here: don't make someone feel like they don't know anything just because they don't know what you know.